locust-iothub/requirements.txt
locust
paho-mqtt
azure-core
azure-identity
azure-keyvault-secrets
azure-iot-hub
azure-data-tables
azure-eventhub
azure-storage-queue